support continuation after failure

+++ 2020-FDL-X-Geo/preprocess.py
@@ -25,13 +25,13 @@ import pickle
 
 # Preprocessing parameters
 past_omni_length = 120
-lag = 0
-yearlist = list(np.arange(2010,2012).astype(int))
+lag = 30
+yearlist = list(np.arange(2016,2019).astype(int))
 targets = ["dbe_nez", "dbn_nez"]
 future_length = 1
 nmax = 20
 num_workers = 16
-output_folder = './processed_data_all_years'
+output_folder = './processed_data_base'
 
 
 class PreprocessData():
@@ -211,29 +211,34 @@ class PreprocessData():
 if not os.path.exists(output_folder):
     os.mkdir(output_folder)
 
-supermag_data = SuperMAGIAGADataset(*get_iaga_data_as_list(base="data_local/iaga/",year=yearlist))
-omni_data = OMNIDataset(get_omni_data("data_local/omni/sw_data.h5", year=yearlist))
+if not os.path.exists(os.path.join(output_folder, 'scalers.p')):
 
-train_idx,test_idx,val_idx,wiemer_idx = generate_indices(base="data_local/iaga/",year=yearlist,
-                                                        LENGTH=past_omni_length,LAG=lag,
-                                                        omni_path="data_local/omni/sw_data.h5",
-                                                        weimer_path="data_local/weimer/")
-train_idx = np.asarray(train_idx)
+    supermag_data = SuperMAGIAGADataset(*get_iaga_data_as_list(base="data_local/iaga/",year=yearlist))
+    omni_data = OMNIDataset(get_omni_data("data_local/omni/sw_data.h5", year=yearlist))
 
-train_ds_all_years = PreprocessData(supermag_data,omni_data,train_idx,
-                f107_dataset="data_local/f107.npz",targets=targets,past_omni_length=past_omni_length,
-                past_supermag_length=1,future_length=future_length,lag=lag,zero_omni=False,
-                zero_supermag=False,scaler=None,training_batch=True,nmax=nmax, num_workers=num_workers)
+    train_idx,test_idx,val_idx,wiemer_idx = generate_indices(base="data_local/iaga/",year=yearlist,
+                                                            LENGTH=past_omni_length,LAG=lag,
+                                                            omni_path="data_local/omni/sw_data.h5",
+                                                            weimer_path="data_local/weimer/")
+    train_idx = np.asarray(train_idx)
 
-scaler = train_ds_all_years.scaler
-with open(os.path.join(output_folder, f'supermag_features.p'), 'wb') as f:
+    train_ds_all_years = PreprocessData(supermag_data,omni_data,train_idx,
+                    f107_dataset="data_local/f107.npz",targets=targets,past_omni_length=past_omni_length,
+                    past_supermag_length=1,future_length=future_length,lag=lag,zero_omni=False,
+                    zero_supermag=False,scaler=None,training_batch=True,nmax=nmax, num_workers=num_workers)
+
+    scaler = train_ds_all_years.scaler
+    with open(os.path.join(output_folder, f'supermag_features.p'), 'wb') as f:
         pickle.dump(train_ds_all_years.supermag_features, f, pickle.HIGHEST_PROTOCOL)
 
-with open(os.path.join(output_folder, f'omni_features.p'), 'wb') as f:
-    pickle.dump(train_ds_all_years.omni_features, f, pickle.HIGHEST_PROTOCOL)
-    
-with open(os.path.join(output_folder, f'scalers.p'), 'wb') as f:
-    pickle.dump(scaler, f, pickle.HIGHEST_PROTOCOL)
+    with open(os.path.join(output_folder, f'omni_features.p'), 'wb') as f:
+        pickle.dump(train_ds_all_years.omni_features, f, pickle.HIGHEST_PROTOCOL)
+
+    with open(os.path.join(output_folder, f'scalers.p'), 'wb') as f:
+        pickle.dump(scaler, f, pickle.HIGHEST_PROTOCOL)
+else:
+    scaler = pickle.load(open(os.path.join(output_folder, 'scalers.p'), 'rb'))
+
     
 max_stations = get_iaga_max_stations(base="data_local/iaga/",yearlist=yearlist)
